TST: Reorganize test_environment_analysis.py.
