+++ tests/integration/test_environment_analysis.py
@@ -1,0 +1,59 @@
+import copy
+import os
+from unittest.mock import patch
+
+import matplotlib as plt
+import pytest
+
+from rocketpy.tools import import_optional_dependency
+
+plt.rcParams.update({"figure.max_open_warning": 0})
+
+
+@pytest.mark.slow
+@patch("matplotlib.pyplot.show")
+def test_all_info(mock_show, env_analysis):
+    """Test the EnvironmentAnalysis.all_info() method, which already invokes
+    several other methods. It is a good way to test the whole class in a first view.
+    However, if it fails, it is hard to know which method is failing.
+
+    Parameters
+    ----------
+    env_analysis : rocketpy.EnvironmentAnalysis
+        A simple object of the Environment Analysis class
+
+    Returns
+    -------
+    None
+    """
+    assert env_analysis.info() == None
+    assert env_analysis.all_info() == None
+    assert env_analysis.plots.info() == None
+    os.remove("wind_rose.gif")  # remove the files created by the method
+
+
+@pytest.mark.slow
+@patch("matplotlib.pyplot.show")
+def test_exports(mock_show, env_analysis):
+    """Check the export methods of the EnvironmentAnalysis class. It
+    only checks if the method runs without errors. It does not check if the
+    files are correct, as this would require a lot of work and would be
+    difficult to maintain.
+
+    Parameters
+    ----------
+    env_analysis : EnvironmentAnalysis
+        A simple object of the EnvironmentAnalysis class.
+    """
+
+    assert env_analysis.export_mean_profiles() == None
+    assert env_analysis.save("env_analysis_dict") == None
+
+    env2 = copy.deepcopy(env_analysis)
+    env2.load("env_analysis_dict")
+    assert env2.all_info() == None
+
+    # Delete file created by save method
+    os.remove("env_analysis_dict")
+    os.remove("wind_rose.gif")
+    os.remove("export_env_analysis.json")
