+++ tests/unit/test_environment_analysis.py
@@ -12,28 +12,6 @@ plt.rcParams.update({"figure.max_open_warning": 0})
 
 @pytest.mark.slow
 @patch("matplotlib.pyplot.show")
-def test_all_info(mock_show, env_analysis):
-    """Test the EnvironmentAnalysis.all_info() method, which already invokes
-    several other methods. It is a good way to test the whole class in a first view.
-    However, if it fails, it is hard to know which method is failing.
-
-    Parameters
-    ----------
-    env_analysis : rocketpy.EnvironmentAnalysis
-        A simple object of the Environment Analysis class
-
-    Returns
-    -------
-    None
-    """
-    assert env_analysis.info() == None
-    assert env_analysis.all_info() == None
-    assert env_analysis.plots.info() == None
-    os.remove("wind_rose.gif")  # remove the files created by the method
-
-
-@pytest.mark.slow
-@patch("matplotlib.pyplot.show")
 def test_distribution_plots(mock_show, env_analysis):
     """Tests the distribution plots method of the EnvironmentAnalysis class. It
     only checks if the method runs without errors. It does not check if the
@@ -132,6 +110,34 @@ def test_profile_plots(mock_show, env_analysis):
 
 
 @pytest.mark.slow
+def test_values(env_analysis):
+    """Check the numeric properties of the EnvironmentAnalysis class. It computes
+    a few values and compares them to the expected values. Not all the values are
+    tested since the most of them were already invoke in the previous tests.
+
+    Parameters
+    ----------
+    env_analysis : EnvironmentAnalysis
+        A simple object of the EnvironmentAnalysis class.
+
+    Returns
+    -------
+    None
+    """
+    assert pytest.approx(env_analysis.record_min_surface_wind_speed, 1e-6) == 5.190407
+    assert (
+        pytest.approx(env_analysis.max_average_temperature_at_altitude, 1e-6)
+        == 24.52549
+    )
+    assert (
+        pytest.approx(env_analysis.min_average_temperature_at_altitude, 1e-6)
+        == -63.18178
+    )
+    assert pytest.approx(env_analysis.std_pressure_at_10000ft, 1e-6) == 13.58699
+    assert pytest.approx(env_analysis.std_pressure_at_30000ft, 1e-6) == 38.48947
+
+
+@pytest.mark.slow
 @patch("matplotlib.pyplot.show")
 def test_animation_plots(mock_show, env_analysis):
     """Check the animation plots method of the EnvironmentAnalysis class. It
@@ -160,58 +166,3 @@ def test_animation_plots(mock_show, env_analysis):
         HTML,
     )
     os.remove("wind_rose.gif")  # remove the files created by the method
-
-
-@pytest.mark.slow
-@patch("matplotlib.pyplot.show")
-def test_exports(mock_show, env_analysis):
-    """Check the export methods of the EnvironmentAnalysis class. It
-    only checks if the method runs without errors. It does not check if the
-    files are correct, as this would require a lot of work and would be
-    difficult to maintain.
-
-    Parameters
-    ----------
-    env_analysis : EnvironmentAnalysis
-        A simple object of the EnvironmentAnalysis class.
-    """
-
-    assert env_analysis.export_mean_profiles() == None
-    assert env_analysis.save("env_analysis_dict") == None
-
-    env2 = copy.deepcopy(env_analysis)
-    env2.load("env_analysis_dict")
-    assert env2.all_info() == None
-
-    # Delete file created by save method
-    os.remove("env_analysis_dict")
-    os.remove("wind_rose.gif")
-    os.remove("export_env_analysis.json")
-
-
-@pytest.mark.slow
-def test_values(env_analysis):
-    """Check the numeric properties of the EnvironmentAnalysis class. It computes
-    a few values and compares them to the expected values. Not all the values are
-    tested since the most of them were already invoke in the previous tests.
-
-    Parameters
-    ----------
-    env_analysis : EnvironmentAnalysis
-        A simple object of the EnvironmentAnalysis class.
-
-    Returns
-    -------
-    None
-    """
-    assert pytest.approx(env_analysis.record_min_surface_wind_speed, 1e-6) == 5.190407
-    assert (
-        pytest.approx(env_analysis.max_average_temperature_at_altitude, 1e-6)
-        == 24.52549
-    )
-    assert (
-        pytest.approx(env_analysis.min_average_temperature_at_altitude, 1e-6)
-        == -63.18178
-    )
-    assert pytest.approx(env_analysis.std_pressure_at_10000ft, 1e-6) == 13.58699
-    assert pytest.approx(env_analysis.std_pressure_at_30000ft, 1e-6) == 38.48947
